Fuck microsoft

+++ src/rigidPlayer.hpp
@@ -109,7 +109,10 @@ class RigidPlayer : public RigidBody3D{
         float airtime = 0;
         bool toonjump = true;
 
-        u_int8_t MaxContactReportCount = 4;
+        int MaxContactReportCount = 4;
+        // cant use a u_int here cuss microsoft and there dword or whatever i guess
+        // tilted about it ngl, dont make this go negative cuss somthing bad will prob happen
+
 
 
         float CurrentSpeed;
